Stage 1-7 -
First part of the project has been completed.

This includes getting data via a promise in GitHub, from a fake database.
Getting all initial data, that is users and questions via a fake api.

Once this is created we need to eventually pass these onto our components. In order  to do that, we need to have them in the store. In order to do that, we have action creators for each action.

Then we need a shared file, that will take the data from the api and dispatch actions.
The actions will then "activate" certain reducers, that will perform actions to the state.
All of these actions are combined into one object and these are passed to the createStore function, which essentially will create the store.

In order to pass the store onto the components, we need a provider, where we pass in the store as an argument
<Provider>
<App>
/Provider>

DATA -> Api ->