action creators must have the same name as dispatched actions?
Why is it that the the action creator function, needs to explicitly be passed in exactly the name of the specific object, it is to be passed?
Isn't just a parameter? Meaning that:
in theory, i would expect to do something like this:

then in my shared.js I am explicitly saying that the argument
that is passed to the receiveUsers action creator, is the users object, containing all users.
Thus the action creator must return us the users, irrespective if in the action function, I asked it to return abracadabra. The argument its being passed is still users.

I discovered this by an accidental typo, i typed in the action creator, user in singular, whereas in the shared.js file I was dispatching the action and passed it the action users (plural)

Meaning if i pass in any other argument in the action creator function, that has does NOT have the same name as the object passed to it, when dispatching an action
which in this case was user in the action creator and users was being dispatched from the shared.js file,
it will simply ignore it.

My question is why this is? As well as how come I can pass in here
the ID in loggedInUser?