Creation of Question.js
currently creating question.js component that will be passed the question id, and then with the use of format question, will take the given prop (id) and also the users and questions and the return formatted data. Currently experiencing:  Error: Objects are not valid as a React child (found: object with keys {id, loggedInUser, question, dispatch}). If you meant to render a collection of children, use an array instead.