Answer/Unanswered Questions
so currently I am getting the data from the users,
i check if the id passed to the AnswerUnansweredQuestion component, contains data from the loggedInUser answers, if that is the case then it's considered answered.
However if it hasn't it's considered unanswered.

Currently I am testing this on a single user. Later I will implement this such that it works for the loggedInUser.