Fixing logic in AnsweredUnansweredQuestions
Currently I am working on the answered/ un ansnwered component.

What essentially is happening, is that it is being passed a prop from the Question component. It then checks to see if this question is part of the loggedInUsers property. If it is, it is considered loggedIn, and such a UI will be presented. If not, it will render Not logged In.

This is logic on line, 33.
If the loggedInUserQuestions contains the questionsParsed, I want to render QuestionParsed - which essentially is all answered questions.

However, I cannot see the rendered h1.