WouldYouRather only showing one item, not all
The questions slice of the store is checked up against: the loggedInUser (saraDoe) answers.

If it can see the question slice is a part of the loggedInUsers answers, it means that the loggedInUser answered it and thus will be rendered.

One issue I am seeing is that on line 24 and 35 we are seeing all question ID's correctly, however when comparing them on line 39 and rendering it on line 47.
It only compares the first value, and thereby renders the first value.

The way in which it should be working, is by it comparing every single value, and rendering it.
So essentially approxiametly 6 questions are to be displayed as the user Sara Doe has answered 6.